feat(premises): refactor price fields to use integers instead of decimals
- Updated `price_per_month`, `price_per_sqm`, and `full_sell_price` to use `PositiveBigIntegerField`.
- Modified related models, migrations, and validation logic accordingly.
- Updated API schemas, services, and test cases for integer-based price fields.
- Replaced Decimal usage with int in price-related operations and formatting.

+++ backend/apps/re_objects/schemas.py
@@ -37,9 +37,9 @@ class PremiseListOut(Schema):
     uuid: str  # Публичный идентификатор (UUID)
     building_uuid: str  # UUID здания, к которому относится помещение
     name: str
-    price: Optional[Decimal] = None
-    sale_price: Optional[Decimal] = None
-    rent_price: Optional[Decimal] = None
+    price: Optional[int] = None
+    sale_price: Optional[int] = None
+    rent_price: Optional[int] = None
     address: str
     floor: Optional[int] = None
     area: Decimal
@@ -51,7 +51,7 @@ class PremiseDetailOut(PremiseListOut):
     """Помещение: полная информация для страницы объекта (все поля списка + описание и доп. параметры)."""
 
     description: Optional[str] = None
-    price_per_sqm: Optional[Decimal] = None
+    price_per_sqm: Optional[int] = None
     ceiling_height: Optional[Decimal] = None
     has_windows: bool = True
     has_parking: bool = False
@@ -91,8 +91,8 @@ class BuildingListOut(Schema):
     address: str
     description: str
     geo_point: Optional[BuildingGeoPointOut] = None
-    min_sale_price: Optional[float] = None
-    min_rent_price: Optional[float] = None
+    min_sale_price: Optional[int] = None
+    min_rent_price: Optional[int] = None
     media: list[BaseMediaItemOut]
 
 
@@ -106,8 +106,8 @@ class BuildingDetailOut(Schema):
     geo_point: Optional[BuildingGeoPointOut] = None
     total_floors: Optional[int] = None
     year_built: Optional[int] = None
-    min_sale_price: Optional[float] = None
-    min_rent_price: Optional[float] = None
+    min_sale_price: Optional[int] = None
+    min_rent_price: Optional[int] = None
     media_categories: list[str]
     media: list[BuildingMediaItemOut]
